refactor(team-builder): extract MetaBar cell component
Lv popover + nickname input + gender chip + shiny chip extraction from
IdentityLaneReal. One MetaBar component, two wrappers via variant prop:

- variant="banner" — compact-mode inner row (flex items-center gap-2),
  no Lv popover; s.idBanner wrapper stays in the parent alongside FormChips
- variant="row" — hero-mode wrapper (.heroMetaRow), Lv popover included

Deviation from plan spec: s.idBanner is the parent of both the meta-row
and FormChips. Per the plan's own fallback instruction, only the inner
meta-row was extracted; s.idBanner remains in IdentityLaneReal. Added
genderErrors prop (missing from spec) since the banner variant renders
adjacent FieldErrors for gender validation. The level-popover open state
lives inside MetaBar (cell owns its picker state).
